nrunner: make check_tasks_requirements a core utility function
